+++ src/services/logging.ts
@@ -8,3 +8,7 @@ export function logToDisk(logLine: string) {
 export function openLogDirectory() {
     ipcRenderer.invoke('open-log-dir');
 }
+
+export function logError(error: Error, message: string, info?: string): void {
+    ipcRenderer.invoke('log-error', error, message, info);
+}
