+++ src/utils/logging.ts
@@ -1,5 +1,8 @@
-import { ipcRenderer } from 'electron';
+import log from 'electron-log';
+import { LOG_FILENAME, MAX_LOG_SIZE } from '../config';
 
-export function logError(error: Error, message: string, info?: string): void {
-    ipcRenderer.invoke('log-error', error, message, info);
+export function setupLogging() {
+    log.transports.file.fileName = LOG_FILENAME;
+    log.transports.file.maxSize = MAX_LOG_SIZE;
+    log.transports.console.level = false;
 }
