+++ src/utils/main.ts
@@ -1,9 +1,4 @@
-import {
-    LOG_FILENAME,
-    MAX_LOG_SIZE,
-    PROD_HOST_URL,
-    RENDERER_OUTPUT_DIR,
-} from '../config';
+import { PROD_HOST_URL, RENDERER_OUTPUT_DIR } from '../config';
 import { nativeImage, Tray, app, BrowserWindow, Menu } from 'electron';
 import electronReload from 'electron-reload';
 import serveNextAt from 'next-electron-server';
@@ -14,7 +9,6 @@ import { isDev } from './common';
 import { buildContextMenu, buildMenuBar } from './menu';
 import autoLauncher from '../services/autoLauncher';
 import { getHideDockIconPreference } from '../services/userPreference';
-import log from 'electron-log';
 
 export function handleUpdates(mainWindow: BrowserWindow, tray: Tray) {
     if (!isDev) {
@@ -104,10 +98,4 @@ export async function handleDockIconHideOnAutoLaunch() {
 
 export function enableSharedArrayBufferSupport() {
     app.commandLine.appendSwitch('enable-features', 'SharedArrayBuffer');
-}
-
-export function setupLogging() {
-    log.transports.file.fileName = LOG_FILENAME;
-    log.transports.file.maxSize = MAX_LOG_SIZE;
-    log.transports.console.level = false;
 }
