+++ src/utils/preload.ts
@@ -1,6 +1,4 @@
-import { LOG_FILENAME, MAX_LOG_SIZE } from '../config';
 import { webFrame } from 'electron';
-import log from 'electron-log';
 
 export const fixHotReloadNext12 = () => {
     webFrame.executeJavaScript(`Object.defineProperty(globalThis, 'WebSocket', {
@@ -23,10 +21,4 @@ export function isPlatformMac() {
 
 export function isPlatformWindows() {
     return process.platform === 'win32';
-}
-
-export function setupLogging() {
-    log.transports.file.fileName = LOG_FILENAME;
-    log.transports.file.maxSize = MAX_LOG_SIZE;
-    log.transports.console.level = false;
 }
